requirements.txt
# ===========================================
# LADX - Python Dependencies
# ===========================================
# Install: pip install -r requirements.txt
# ===========================================

# Core AI (OpenRouter uses OpenAI-compatible API)
openai>=1.40.0

# Knowledge Base (RAG)
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
chromadb>=0.5.0
sentence-transformers>=3.0.0

# Web Interface
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
jinja2>=3.1.0
python-multipart>=0.0.9

# Authentication & Security
bcrypt>=4.0.0
pyjwt>=2.8.0

# Database
sqlalchemy>=2.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.27.0
requests>=2.32.0

# Document processing (for knowledge base)
pypdf>=4.0.0
python-docx>=1.1.0
